+++ src/polyfill/htmlelement.js
@@ -147,7 +147,7 @@ Object.defineProperty(HTMLElement.prototype, 'dataset', {
 
 };
 
-//DOM4 mutation methods
+//DOM4 w3c.github.io/dom
 'append' in document.createDocumentFragment() || new function () {
 
     var ELEMENT_NODE = 1,
@@ -199,13 +199,9 @@ Object.defineProperty(HTMLElement.prototype, 'dataset', {
                 this.insertBefore(mutationMacro(arguments), this.firstChild);
             },
 
-            query: function (selector) {
-                return this.querySelector(selector);
-            },
+            querySelector: proto.querySelector,
 
-            queryAll: function (selector) {
-                return this.querySelectorAll(selector);
-            },
+            querySelectorAll: proto.querySelectorAll,
 
             matches: [
                 proto.matchesSelector,
@@ -275,9 +271,13 @@ Object.defineProperty(HTMLElement.prototype, 'dataset', {
     Object.keys(api).forEach(implement);
 
     proto = document.constructor.prototype;
-    ['query', 'queryAll'].forEach(implement);
+    ['querySelector', 'querySelectorAll'].forEach(implement);
 
     proto = document.createDocumentFragment().constructor.prototype;
-    ['append', 'prepend', 'query', 'queryAll', 'matches'].forEach(implement);
+    [
+        'append', 'prepend',
+        'querySelector', 'querySelectorAll',
+        'matches'
+    ].forEach(implement);
 
 };