+++ src/polyfill/ltie10/ie8/head.js
@@ -1,6 +1,6 @@
 
 'head' in document || Object.defineProperty(document, 'head', {
     get: function () {
-        return this.query('head');
+        return document.querySelector('head');
     }
 });